abstract out h2oSetupParseOutput
